Simplified install process
From this commit forward master will always be a stable branch, and all
development will take place in the dev branch. This removes the need for
the install.py script which this commit removes.

This also introduces setup.py entry points. The install instructions
have been updated to reflect this new method of installation.

See #2