Check recent locks before locking a submission
Before locking a submission, check if another TA recently locked that
lab. If so, present the TA with an opportunity to verify that the
student still needs grading. The recent window of time is 10 minutes but
can be easily changed.

It may be good to disable the prompt if the same TA graded it within 10
minutes.

The check for current locks is moved to after the flags and recently
graded popups we can prevent some small chance of a race condition where
both TAs move past the check for locks and are viewing the popups for
flags and recent grades.