Display test results in the grader popup
This option opens the test results for each part in a hierarchical list.
Each part may be expanded to see the score results. Show the score for
the part, or "No Submission" if that part was not submitted.

Because it is hierarchical, we might be able to add more information
inside each test, such as why it failed, or at least have each item be
clickable to view the full test output.