Align the test result scores better
This adds alignment to the scores of the test bench view popup. Each lab
part score is aligned, and then each score within the each part is
vertically aligned as well.

The scores are right-aligned with a padding of 5. While some scores are
of 7 width, this is an exception (like lab 1 with only one test case)
and those don't need special alignment. So in the general case, a
padding of 5 works well and doesn't introduce so much padding that it
looks bad.