Expose recently graded ranges as admin config values
This adds a new Config menu to the admin menu for changing admin/global
options in zygrader. This makes it easier to adjust these values without
needing to make a new zygrader release. This does NOT change the default
values from what they were in previous releases.

The class code menu was also moved to the new Config menu similar to the
class code override in the user preferences.