Increase accuracy of reported scores with compile errors
When a submission failed to compile, the score would show as 0/0. This
improves the results by showing the denominator of the fraction by
reading the max score of the test case.

There is still one issue though. If a student did not submit a part of
an assignment there is no data from zyBooks on the max score, so the
aggregate score will be inaccurate and possibly show 100% when it is in
fact not 100%.